+++ src/modules/Bill/components/SingleBill/CosponsorDialog/index.tsx
@@ -8,8 +8,8 @@ import { Bill } from '@/modules/Bill/classes/Bill'
 import CloseIcon from '@mui/icons-material/Close'
 import { USTWTheme } from '@/common/lib/mui/theme'
 import { Grid2, useTheme } from '@mui/material'
-import DialogFilter from '@/modules/Bill/components/SingleBill/DialogFilter'
-import useDialogFilter from '@/modules/Bill/components/SingleBill/useDialogFilter'
+import DialogFilter from '@/modules/Bill/components/SingleBill/CosponsorDialog/DialogFilter'
+import useDialogFilter from '@/modules/Bill/components/SingleBill/CosponsorDialog/useDialogFilter'
 import CosponsorTable from '@/modules/Bill/components/SingleBill/CosponsorDialog/CosponsorTable'
 import { useMemo } from 'react'
 import { createFilterCategories } from '@/modules/Bill/components/SingleBill/CosponsorDialog/utils'
